@@ -1,5 +1,5 @@
-import {Component, ElementRef, Input, NgZone, OnDestroy, OnInit, ViewChild} from '@angular/core';
-import {Tile} from '../maze.component';
+import {Component, ElementRef, Input, OnInit, ViewChild} from '@angular/core';
+import {Tile} from '../prim.service';
 
 @Component({
   selector: 'app-canvas',
@@ -10,29 +10,26 @@ export class CanvasComponent implements OnInit {
   @Input() width: number;
   @Input() height: number;
   @Input() scaleFactor: number;
+  @Input() timeout: number;
   @Input() mapMemory: Array<Tile>;
 
-  @ViewChild('canvas') canvas: ElementRef;
+  @ViewChild('canvas', {static: true}) canvas: ElementRef;
 
   @Input() running: boolean;
 
+  // Are actually counters for the same: the current place in the memory map.
+  // But both are needed, because of the timeout of the paint action:
+  // When painting is run (after the timeout) the memoryCounter has already reached its end
   private memoryCounter = 0;
-  private timeoutCounter = 1;
   private tileCounter = 0;
+
+  private timeoutCounter = 0;
 
   constructor() {
   }
 
   ngOnInit() {
 
-  }
-
-  private paintMap() {
-    if (this.mapMemory.length == 0) {
-      return;
-    }
-    let ctx = this.canvas.nativeElement.getContext('2d');
-    setTimeout(() => this.paintTile(ctx), 25 * this.timeoutCounter);
   }
 
   private paintTile(ctx) {
@@ -44,20 +41,14 @@ export class CanvasComponent implements OnInit {
   }
 
   play() {
-    this.reset();
-    for (this.memoryCounter = 0; this.memoryCounter < this.mapMemory.length; this.memoryCounter++) {
-      this.paintMap();
-      this.timeoutCounter++;
-    }
-  }
-
-  next() {
-    if (this.memoryCounter == this.mapMemory.length - 1) {
+    if (this.mapMemory.length == 0) {
       return;
     }
-    this.memoryCounter++;
+    this.reset();
     let ctx = this.canvas.nativeElement.getContext('2d');
-    this.paintTile(ctx);
+    for (this.memoryCounter = 0; this.memoryCounter < this.mapMemory.length; this.memoryCounter++) {
+      setTimeout(() => this.paintTile(ctx), this.timeout * this.timeoutCounter++);
+    }
   }
 
   reset() {
@@ -67,14 +58,5 @@ export class CanvasComponent implements OnInit {
     let ctx = this.canvas.nativeElement.getContext('2d');
     ctx.fillStyle = '#979797';
     ctx.fillRect(0, 0, this.width, this.height);
-  }
-
-  previous() {
-    if (this.memoryCounter <= 0) {
-      return;
-    }
-    this.memoryCounter--;
-    let ctx = this.canvas.nativeElement.getContext('2d');
-    this.paintTile(ctx);
   }
 }
